+++ docs/scripts/routes/components/dataview.js
@@ -1,8 +1,13 @@
+const Card = () => import('../../views/components/card');
 const List = () => import('../../views/components/list');
 const Table = () => import('../../views/components/table');
 const Pagination = () => import('../../views/components/pagination');
 
 const dataviewRoutes = [{
+  path: 'card',
+  name: 'components.card',
+  component: Card
+}, {
   path: 'list',
   name: 'components.list',
   component: List