+++ src/scripts/components/card.js
@@ -1,5 +1,7 @@
 import UiCard from './card/card';
+import UiCardHead from './card/card-head';
 import UiCardTitle from './card/card-title';
+import UiCardSubtitle from './card/card-subtitle';
 import UiCardMedia from './card/card-media';
 import UiCardBody from './card/card-body';
 import UiCardActions from './card/card-actions';
@@ -7,7 +9,9 @@ import UiCardMenu from './card/card-menu';
 
 const components = {
   UiCard,
+  UiCardHead,
   UiCardTitle,
+  UiCardSubtitle,
   UiCardMedia,
   UiCardBody,
   UiCardActions,