+++ src/scripts/components/card/card.vue
@@ -11,7 +11,7 @@ export default {
   name: 'ui-card',
   props: {
     // ui attributes
-    dp: {
+    shadow: {
       type: [Number, String],
       default: 2
     }
@@ -19,7 +19,7 @@ export default {
   computed: {
     shadowClassName() {
       // Assigns variable shadow depths (2, 3, 4, 6, 8, or 16) to card
-      return (SHADOW_DEPTHS.indexOf(+this.dp) > -1) ? `mdl-shadow--${this.dp}dp` : '';
+      return (SHADOW_DEPTHS.indexOf(+this.shadow) > -1) ? `mdl-shadow--${this.shadow}dp` : '';
     },
     className() {
       return [